edge-dev-ai-agent/requirements.txt
# EdgeDev AI Agent Requirements

# Core Dependencies
anthropic>=0.18.0
openai>=1.0.0
httpx>=0.25.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0

# MCP Integration
mcp>=0.9.0

# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0
jinja2>=3.1.2
python-multipart>=0.0.6

# Data Science
pandas>=2.0.0
numpy>=1.24.0

# Archon Integration
pydantic>=2.0.0
pydantic-ai>=0.0.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3
